Patched unit test to account for validation bug
In #2bf635e, a bug was missed by the `Map_ComplexObject_ReturnsFlattenedTopic` unit test since the `Contact` content type had an `Email` attribute. As such, when the validator was supposed to be validating e.g. `AlternateEmail`, it was _actually_ validating `Email`, and passing. In a real-world scenario, however, this won't usually be the case, and defeats the benefit of validating the model against the content type.

To account for this, the generic `Email` attribute was removed, and a new `Name` attribute was introduced. This clutters up the unit test and the model a bit by requiring two separate objects (`ContactTopicBindingModel` and `EmailTopicBindingModel`) but this effectively tests the same conditions while additionally validating that the `AttributePrefix` context is being properly accounted for as part of that validation.