Moved to new `Maintenance` folder
By convention, a SQL Database Project organized objects into folders, such as `Stored Procedures`. This is a best practice, and one we generally follow. In addition, however, we have a lot of maintenance stored procedures which aren't usually needed, and it would be nice to move them out of the way. This isn't possible in SQL Server itself, at least without using schemas—which we may reevaluate in the future. We can, however, put them in any local folder we want. With this in mind, we've created a new `Maintenance` folder and moved the internal utilities and maintenance stored procedures to it.