Introduce new `DeleteConsecutiveAttributes` maintenance procedure
In older versions of the OnTopic Library, new attribute versions could be saved even if their value had not changed. This isn't necessary in terms of the OnTopic library, which will take whatever the latest version of an attribute is when loading the topic graph (or restoring a topic). As such, these _consecutive duplicates_ only add clutter to the database, and potentially slow down some queries by introducing more data to sort through. The `DeleteConsecutiveAttributes` identifies and deletes these from the database, thus making it more efficient.