Introduce new `DeleteConsecutiveExtendedAttributes` maintenance procedure
In older versions of the OnTopic Library, new extended attribute versions were _always_ written, even if their composite XML value had not changed. This isn't necessary in terms of the OnTopic library, which will take whatever the latest version of an attribute is when loading the topic graph (or restoring a topic). As such, these _consecutive duplicates_ only add clutter to the database, and potentially slow down some queries by introducing more data to sort through. This is especially concerning for the extended attributes since they use large XML fields. The `DeleteConsecutiveExtendedAttributes` identifies and deletes these from the database, thus making it more efficient. This corresponds to #fc3e27a, which does the same for indexed attributes.