Introduce new `ConsolidateVersions` maintenance procedure
Currently, a version gets saved every time a value is changed. That potentially creates a lot of records over time—far more than the five historical versions we surface to the OnTopic editor. The resolution of those versions becomes impractical at some point: If a topic was edited ten times in a week five years ago, we don't likely care about each of those versions. The `ConsolidateVersions` stored procedure solves this by allowing _all_ versions within a given date range to be consolidated into a single, new version. This reduces the granularity. This can be used to consolidate all topics older than a specific date—or all topics _within_ a date _range_. The latter, for instance, could be used to compress each year down to a single version.