Move stored procedures to new `Stored Procedures` folder
Originally, since `Maintenance` was just intended as an internal organizational structure, the stored procedures were just placed in the root. Now, since we're intending to formalize this by turning it into a different schema, it makes sense to move these down into a `Stored Procedures` directory, as per the convention for SQL Server projects in Visual Studio.