Introduced `SqlCommandExtensions`
Refactored the `AddSqlParameter()` helper function(s) as extension methods off of the `SqlCommand` object, and added strongly typed overloads so that the `SqlDbType` doesn't need to be passed along. Also added a new `AddOutputParameter()` method so that `ParameterDirection` doesn't need to be specified. Finally, removed `sqlLength` parameter since it wasn't being used, and cleaned up the final `AddParameter()` overload to be more succinct by e.g. using pattern matching with a switch statement.