Introduced `SqlDataReaderExtensions`
Established a series of strongly typed extension methods on `SqlDataReader` for retrieving values. The automatically validate that the column exists, lookup the column name using `GetOrdinal()`, and use the out-of-the-box strongly typed methods (e.g., GetBoolean()`) to retrieve the value in a native data format (e.g., `bool`). This approach doesn't currently support nullable values due to the `T?` challenges—but we also don't currently have a need to read values that any nullable columns, so that isn't yet a limitation. Obviously, we may need to expand these in the future if it becomes an issue.