Introduced `TopicRepositoryBase.GetAttributes()`
The new `GetAttributes()` utility method simplifies the retrieval of attributes by optionally allowing them to be filtered by either `IsExtendable` and `IsDirty`. To filter by `IsExtendable`, the results must be cross-referenced against the local `ContentTypeDescriptor`s via `GetContentTypeDescriptors()`. This logic was previously part of `SqlTopicRepository`, but there's nothing about it which is _specific_ to SQL, and therefore there's value to centralizing it, not only to support potential future `ITopicRepository` implementations, but _also_ to help simplify the code in the already bloated `SqlTopicRepository`. This provides a collection useful for determining the indexed attributes (`!IsExtendedAttribute && IsDirty`) as well as the extended attributes (`IsExtendedAttribute`).