Removed workaround for `CA1062` false positives
With dotnet/roslyn-analyzers#2369 resolved, we are now able to reintroduce `String.IsNullOrEmpty(value)` for 'null` checks without worrying about false positives on `CA1062` regarding lack of validation. Previously, to avoid tripping this issue, we had been instead using e.g. `value is null || value.Length.Equals(0)`.

That said, this _still_ produces a false positive with `CS8601` and `CS8604` in Roslyn's static flow analysis (i.e., "possibly null reference assignment" and "possible null reference argument for parameter…"). As such, we're unable yet to remove the workaround for cases where the reference is being assigned to a non-nullable variable or parameter, such as in `Contract.Requires<T>()`.

The latter is only an issue because we continue to target .NET Standard 2.0 for backward compatibility with the .NET Framework, but .NET Standard 2.0 doesn't support nullability context. As such, until we decide to only target .NET Standard 2.1+, those cases will likely need to be maintained.