Disable `CA1031` warnings in known cases
`CA1031` pertains to catching general exceptions, such as `ArgumentException` or `InvalidOperationException`, in preference for more specific exceptions. In these cases, howeer, we don't control the exceptions being thrown by the base class library, and thus these are already the most specific exceptions we can catch. To mitigate that, we're wrapping these `catch` statements in a `#pragma` directive to disable these particular warnings. It's not pretty, but it works. It's worth noting that we could put in more code to detect the conditions that lead to these exceptions—such as a duplicate key exception on a `KeyedCollection` leading to an `ArgumentException`—but this would require a fair amount of extra code for what is effectively a corner case.