Introduced support for implicit default
In #dde0ade, I noted that we didn't include support for `DBNull`, but that wasn't a concern since none of the fields were nullable. That was incorrect. There is one field that is nullable: `ParentId`. The field this is drawn from (`AttributeValue`) is not nullable, but if that value doesn't exist then it will be returned as `null`. To mitigate this, I've added a `defaultValue` parameter which defines an implicit default for each value if the actual value can't be identified. Normally, we'd use `default(T)` for this, but that isn't practical here since `T` could be _either_ a `class` _or_ a `struct`, and there isn't a way to setup a generic condition that supports both of those. Just as well, however, as we don't always want the default value; for integers, for instance, we want them to return `-1` to differentiate them from e.g. possibly valid entity references. While I was at it, I added implicit defaults for all supported types (e.g., `bool`, `DateTime`, &c.), even though we _should_ only need them for integers right now.