Introduced new `GetVersion()` extension
When we introduced the `GetDateTime()` extension method in #dde0ade, we inadvertantly reduced the precision of the `DateTime` value returned. This is a problem since this is used for `Version` (e.g., when setting the `VersionHistory`) and that's a place where the precision must match the database exactly. Unfortunately, the logic for this is a bit different than it is for other types, since it's the interstitial string conversion that's introducing the problem here. To mitigate that, we must duplicate some of the logic from the private `GetValue<T>()` in order to explicitly avoid that conversion, and instead pull the value explicitly using the `GetDateTime()` method. This process won't actually work with other types of `DateTime` values, such as those encoded as a string in the `Attributes` table. As such, to differentiate these, this is being introduced as `GetVersion()`.