Remove unused extension methods
When I introduced the `SqlDataReader` extension methods in #dde0ade, I added a few that _might_ be needed—and, indeed, might _still_ be needed in the future—but aren't needed _yet_. Since this is a private library intended only to support the `SqlTopicRepository`, there's no real benefit to maintaining private methods that aren't being used. And, in fact, given _how_ they're being used, it's unlikely (but not impossible) that we'll need them given the current approach to the SQL schema. (I.e., since most attributes are stored as strings in both the database as well as the topic library, there isn't a need to use e.g. `GetBoolean()`, unless we add a new metadata column that's a `BIT`.)