Introduced new `GetTopicId()` extension method
The logic for retrieving a `TopicID` is a bit different than retrieving an attribute that happens to be an integer since it doesn't require a conversion, and can be retrieved directly as an integer. By sharing `GetInteger()` for both use cases, we were requiring that the `TopicID` be converted to a string and then back to an integer for _every single topic_—and sometimes more frequently (for relationships). By introducing the `GetTopicId()` extension, we prevent this unnecessary overhead. It's a relatively trivial improvement, but it doesn't require much code, and it also provides a bit more specific semantics (i.e., `GetTopicId()` instead of `GetInteger("TopicID")`.