Removed `GetValue()` private helper method
The initial implementation of the extension methods (#f25a063) was a bit too cute with an attempt to centralize the logic. With subsequent efforts to refine this logic (#28948f7, #47ef38d, #adcf6a1, #294b59d) the cracks started to show in this approach. At this point, it is either not used (e.g., for `GetVersion()`), or is providing unnecessary overhead (e.g., by converting `INT` fields to a string value and then back to an `int`).  As such, the private `GetValue()` helper was removed, and the final two extension methods were converted to using local methods. The only one that looks even remotely like the original `GetValue()` helper is `GetInteger()`, but it is considerably simpler since the