Fixed validation of `ParentId` for recursive `Save()`
When doing a recursive `Save()`, the `ParentID` attribute of each child topic was being validated to ensure that it had correctly been set to a positive number—i.e., that the parent's `Id` had been properly set from the database. There were two issues with this. First, we really only need to check that once, when the `topic.Id` is set, not for each child. Second, it was being checked _prior_ to the `ParentID` being set on the child. While I was at it, I set the `ParentID` based on `topic.Id` not `returnVal`; these will be the same, but the semantics are a bit clearer.