Removed unnecessary `Assume()`
There were a number of unnecessary `Assume()` checks. These are necessary when checking fields that could be `null`. Since the database columns these are checking are not nullable, however, this condition should never happen. And, if it does, the `SqlDataReader` extension methods will likely throw an exception before these `Assume()` methods are ever called, since most of the `SqlDataReader` extensions are (by design) not aware of `DBNull`.