Standardized identifiers of `topicId`, `attributeKey`, and `attributeValue` variables
Previously, `topicId` was either `id` or `sourceTopicId`; `attributeKey` was `name`; and `attributeValue` was `value`. To ensure these are more consistent with the database values they represent, these have been changed to `topicId`, `attributeKey`, and `attributeValue` respectively. The most notable of these was `name`, since it didn't map to any standard identifiers. The others are a bit more debatable, since they are otherwise consistent with the object model. But being explicit when working with the SQL interop is useful.