Consolidated the content type validation logic
Previously, the `contentType` (now `contentTypeDescriptor`) was defined in one block, but validated in a later block. These are now brought together. In addition, the `GetContentTypeDescriptors()` collection is confirmed to contain the content type before retrieving it. Otherwise, the validation will never fire, and instead a less-intuitive `IndexOutOfRangeException` will be thrown.

It's worth noting that we no longer _need_ the `ContentTypeDescriptor` here. Instead, this is needed by the new `GetAttributes()` and `GetNullableAttributes()` methods on `TopicRepositoryBase` (see #6662f07, #2237a02), which call `GetContentTypes()` on their own. To reduce the number of times this needs to be looked up and validated, it may make more sense to just relay this to `GetAttributes()` and `GetNullableAttributes()` in the future; we'll need to consider that.