Moved `Load()` helpers to `SqlDataReaderExtensions`
Most of the `Load()` helper functions operate off of individual records returned from `SqlDataReader`, and can thus be implemented as extensions. Better yet, these can be kept as `private` extensions, with a single internal `LoadTopicGraph()` extension as the entry point, thus hiding much of the internal complexity from `SqlTopicRepository`. This doesn't really reduce the amount of code or complexity involved in `SqlTopicRepository`, but it does help reorganize the code, and keep `SqlTopicRepository` focused on the higher level logic, instead of the parsing of individual records in order to populate the topic graph.