Moved parameters outside of `try/catch` blocks
There doesn't appear to be a strong benefit to wrapping the `SqlCommand` and `Parameters` in the `try/catch` statement, and by moving them to the main section, we a) reduce nesting, b) are able to reduce comment blocks (since the code within is much smaller and easier to read), and c) keep related code close to one another (i.e., the definition of the `SqlConnection`, `SqlCommand`, and `SqlParameter` objects).

Note: I have read concerns on e.g. Stack Overflow that if a connection cannot be established with the server or the database, then the exception will be thrown at the point where the `SqlConnection` or `SqlCommand` were initialized, not where the connection was `.Open()`ed, or the command was executed. I wasn't able to reproduce this, however; missing server, missing databases, and missing objects all happened on one of those two methods, and were thus caught by the `catch` statement—and, perhaps more importantly, the `finally` statement. As such, there doesn't appear to be any benefit to initializing the `SqlConnection`, `SqlCommand`, or `SqlParameter` objects outside of the `try/catch` blocks.