Put in place condition for `PageLevelNavigation`
Previously, if a page didn't include `PageLevelNavigation`, the partial view would fail due to an attempt to iterate a `null` object. This fixes that issue by wrapping that loop in a condition. This allows a far greater number of pages to be previewed in the local `Host` project.

That said, when connecting to customer databases for testing, this will still fail to recognize any pages associated with proprietary content types, since those won't have local views associated with them.