Fix logic mixup with newly introduced `procedureName` variable
When I introduced the new `procedureName` variable to track whether the `Save()` was calling `CreateTopic` or `UpdateTopic` (#ac6c2ba6), I accidentally got the logic wrong. As a result, `CreateTopic` was being called for new topics, and `UpdateTopic` for existing topics—which caused a mismatch with the parameters. Oops. To fix this, I now use the recently introduced `isNew` variable (#36ca32d), which is a bit more semantically intuitive than `topic.Id > 0`.