Updated `version` to use `SqlDateTime`
The `DateTime` data type in SQL has a lower precision than the `DateTime` data type in C#. As a result, when storing the `version` in `topic.VersionHistory`, a different version was getting saved than in the database. In practice, this resulted in an inability to roll back versions that had been created since the last application reset. (Previous versions worked, as they were loaded with the lower-precision value from the database.) This is mitigated by instead initializing the `version` variable as a `SqlDateTime` object, which uses the same precision as the database.

While I was at it, I move the `VersionHistory` insert into the `try/catch` block so that a `version` won't be prepended if there was an error during the `Save()` operation. (In practice, this is unlikely, and we haven't had any reports of this being a concern, but it's an easy fix to make while modifying the code.)