Marked private helper methods as `[Obsolete]`
The `SitemapController` includes two methods which should not be marked as `public` as they're only intended for use within the `SitemapController` itself. In OnTopic Library 5.0, we'll move these to `private`, thus effectively removing them from the public API. Nobody should be using them, but this warning will help be sure of that. While I was at it, sorted and removed unnecessary `using` statements.