Removed placeholder comment for import/export functionality
For six years, we've been carrying around this commented out placeholder for a `Load()` method that would handle importing topics from an `XmlDocument` in order to support import capabilities. We shouldn't maintain commented out placeholder code like this to begin with. But, in addition, it's no longer relevant as we've implemented a import capabilities via the new **OnTopic-Data-Transfer** library—which is also a much more elegant and flexible approach. Good riddance!