Marked unused event handlers as `[Obsolete]`
For a decade, we've continued to maintain the `DeleteEvent`, `MoveEvent`, and `RenameEvent` on `ITopicRepository` and its predecessors. No implementation has actually ever actually used them, however. There may be a good use case for these (I can imagine several) but the fact that we have never ended up utilizing them makes their value unconvincing.

Further, they're not currently covered by unit tests, which means we're not confident they even work as expected.

Ultimately, if we want to implement events for a project, we likely want far more extensive event coverage (e.g., for other services, for before and after states, &c.) and should revisit this as a more deliberate effort.