Established new `TopicQueryingTests`
Previously, unit tests for the `OnTopic.Querying.TopicExtensions` class were placed in the `TopicTest` class. This made good sense as we only had a single unit test and it related to the `Topic` class. As we prepare to expand the `TopicExtensions` class, however, there's value to separating these out so that the extension methods are independently validated from the main `Topic` class.

To kick this off, the existing `FindAllByAttribute_ReturnsCorrectTopics` unit test is being moved from `TopicTest` to `TopicQueryingTest`. In the near future, additional tests will be introduced, as new features are introduced.