Introduced new `GetRootTopic()` extension method
Originally, the `OnTopic.Querying` class was focused on querying the current `Topic` and its descendants. Increasingly, however, we're identifying needs to work with the entire in-memory topic graph starting from an arbitrary position within the tree. This is particularly useful when it's difficult or inappropriate to inject an `ITopicRepository` such that we have access from the root.

To start, the `GetRootTopic()` extension now allows us to easily access the root topic of any graph starting from any arbitrary point. It's a really simple recrusive function.

This includes a unit test for validating the functionality.