Introduced new `GetTopicByUniqueKey()` extension method
Originally, the `OnTopic.Querying` class was focused on querying the current `Topic` and its descendants. Increasingly, however, we're identifying needs to work with the entire in-memory topic graph starting from an arbitrary position within the tree. This is particularly useful when it's difficult or inappropriate to inject an `ITopicRepository` such that we have access from the root.

As part of this, we're introducing the new `GetTopicByUniqueKey()` extension method, which allows any topic to be retrieved based on its absolute (or root relative) path in the topic graph, starting from any arbitrary position. This is useful for e.g. populating strongly typed topic references for `Relationships`, `IncomingRelationships`, and `DerivedTopic` , as well as metadata references for e.g. the associated `ContentTypeDescriptor`, even though it may not otherwise be appropriate to maintain a dependency on `ITopicRepository`.

Includes unit tests for validating the functionality.