Implemented new `GetTopicByUniqueKey()` extension
The new `GetTopicByUniqueKey()` extension takes the functionality previously implemented in the private `GetTopic()` method and makes it broadly available for other aspects of the application. This allows us to greatly simplify the code in `CachedTopicRepository` by replacing `GetTopic()` entirely with a simple call to `GetByUniqueKey()`.

Ironically, this new approach should actually be _faster_, and certainly requires _less code_, even despite having to start from an arbitrary place within the topic graph. This is because of a more economical approach to how the tree is traversed by simply breaking the `uniqueKey` into an array and looping over that array as part of the traversal.