Established new `TopicRepositoryBaseTest`
Previously, unit tests for the `OnTopic.Repositories.TopicRepositoryBase` class were placed in the `ITopicRepositoryTest` class. This made reasonable sense as the `TopicRepositoryBase` is an `abstract` class, and almost all of the functionality needs to be tested through a concrete implementation such as `StubTopicRepository` anyway.

As we prepare to expand the `TopicRepositoryBase` class, however, there's value to separating these out so that the methods can be independently validated. These are still dependent on a concrete implementation, but that work can be done through simple pass-through methods.

As some tests will still benefit from having prepackaged data to work with, we'll continue to operate against the `StubTopicRepository`—as opposed to the simpler `DummyTopicRepository`—though for many tests we'll just be relying on minimal pass-through methods. In fact, for most cases, we'll be exposing "proxy" methods to publicly expose the underlying `protected` methods with no intermediary business logic.