Introduced `IsExtendedAttribute` support for `CreateFakeData()`
Previously, while the `StubTopicRepository` introduced a number of common `ContentTypeDescriptor` and `AttributeDescriptor` definitions, it didn't set the `IsExtendedAttribute` definition.

In preparation for adding unit tests for `GetAttributes()`, we're introducing support for this by extending the local helper function `addAttribute()`, and factoring that into the definitions.

This should remain backwards compatible since no existing tests are relying on or referencing these data.