Introduced unit tests for `GetAttributes()`
In a previous commit, we introduced a new `GetAttributes()` method to `TopicRepositoryBase` (6662f07), but neglected to include unit test coverage. As we prepare to extend this functionality, we're including the missing unit tests to ensure that we're not breaking existing functionality in the process.

This includes a "proxy" method on `StubTopicRepository` to publicly expose the `protected` `GetAttributes()` method via a `GetAttributesProxy()` passthrough.