Update `GetContentTypeDescriptors()` cache on `Save()`, `Delete()`
While many updates to the configuration metadata are made available in real time, some—such as adding or deleting `ContentTypeDescriptor`s—are not. This is because the `TopicRepositoryBase.GetContentTypeDecriptors()` method caches the list of content types so that it doesn't need to reload it from the persistence store each time its called.

To mitigate this, we can dynamically update the `GetContentTypeDescriptors()` cache when we call `Save()` or `Delete()`. Specifically, if `Save()` is called with a `ContentTypeDescriptor` and that content type is new, then it will be added to the local `_contentTypeDescriptors` collection. Similarly, if `Delete()` is called with a `ContentTypeDescriptor` then that content type—along with all descendent content types—will be removed from the _contentTypeDescriptors` collection.

This satisfies the requirements of the newly proposed feature enhancement: Update content type cache when adding or removing content types (#16).