Update `AttributeDescriptors` cache on `Save()`, `Delete()`, and `Move()`
While many updates to the configuration metadata are made available in real time, some—such as adding or deleting `AttributeDescriptor`s—are not. This is because the `ContentTypeDecriptors.AttributeDescriptors` property caches the list of attributes so that it doesn't need to recalculate them each time its called.

Specifically, the property not only pulls in `AttributeDescriptor` values from the local `Attributes` collection, but also from all ascendent `Attributes` collections of parent `ContentTypeDescriptor` object. This reflects the fact that attributes are inherited from parent content types.

To mitigate this, I've introduced a new `internal` `ResetAttributeDescriptors()` method on `ContentTypeDescriptor`. This simply sets the local backing field for the `AttributeDescriptors` collection to `null`, and then recursively does the same for any `ContentTypeDescriptor` objects underneath it.

Critically, it does not repopulate the cache at this point; instead, that will be done the first time the `AttributeDescriptors` collection is called. This is important, as otherwise that process would be repeated multiple times when doing a recursive `Save()`—as is common when using the **OnTopic Data Exchange** library. By simply nullifying the field, we can safely call this method multiple times with minimal overhead; the caches will then be repopulated later, each time the respective collection is referenced.

Similar to how we update the `GetContentTypeDescriptors()` cache (a9ea634), this his new `ResetAttributeDescriptors()` method is called from `Save()`, `Delete()`, and `Move()` any time they are called with an `AttributeDescriptor`. More specifically, this occurs on `Save()` if the topic is new, and on `Move()` if the parent topic is different than the target topic.

This satisfies the requirements of the newly proposed feature enhancement: Update attribute descriptors when adding or removing an attribute (#17).