Support saving arbitrary attributes
When an `ITopicRepository` implementation calls `GetAttributes()`, they will now receive arbitrary attributes—i.e., attributes that aren't specified in the `ContentTypeDescriptor` schema, but are defined in the `AttributeValueCollection`. This allows those arbitrary values to be programmatically defined and `Save()`d using the **OnTopic Library**. Previously, this functionality was exclusively available via the SQL database.

Arbitrary attributes present some ambiguity. One source of ambiguity is whether they should be stored as _indexed_ or _extended_ attributes. Currently, most arbitrary attributes defined in the database are stored as _indexed_ attributes, which are easier to update and lookup via SQL. They also tend to be short, so this is an optimal format for them. That said, an arbitrary attribute _could_ be longer than 255—the limit for indexed attributes.

Given this, the `GetAttributes()` will return an arbitrary attribute as _indexed_ if the length is 255 characters or less, and as _extended_ if it's greater than 255 characters. As part of this, I've also made the condtion for returning an attribute more concise.

Finally, this includes a couple of unit tests for validating that arbitrary values are correctly being returned as either indexed or extended attributes via `GetAttributes()`.

This supports Programmatic Support for Arbitrary Attributes (#19)